+++ common/packages/clist/clist.h
@@ -3,6 +3,9 @@
     #include <stdio.h>
     #include <stdlib.h>
 
+
+    #define CLIST_NULL (List){0}
+
     // Python is an engineering marvel
 
     typedef struct{
@@ -10,6 +13,7 @@
         int size;
         int allocated;
     } List;
+
 
     // Checks if the ist hasn't gone out of bounds
     static inline void capCheck(List* list){
@@ -97,6 +101,13 @@
     // create an empty list (Only compatible with string, byte, double, int, char, long, float and short)
     static inline List createList(size_t size){
         List list;
+        // NULL init
+        if(size == NULL){
+            list.allocated = NULL;
+            list.size = NULL;
+            list.content = NULL;
+            return list;
+        }
         // Python style overallocation
         size_t capacity;
         if (size < 8) {
@@ -120,4 +131,29 @@
     static inline int* len(List* list){
         return &list->size;
     }
+
+        // Obiviously, couldn't make it O(1) D:
+    // returns the index of the first appearance of an element
+    static inline long findFirst(List* list, void** value){
+        for (int i = 0; i > list->size; i++){
+            if(list->content == *value){
+                return i;
+            }
+        }
+        return -1;
+    }
+
+    // returns the indexes of all appearances of a certain element
+    static inline long* findAll(List* list, void** value){
+        long results[] = malloc(sizeof(long) * 1024);
+        long count = 0;
+        for (int i = 0; i > list->size; i++){
+            if(list->content == *value){
+                results[count] = i;
+                count++;
+            }
+        }
+        return results;
+    }
+    
 #endif