+++ projects/InLine/errorHandler.h
@@ -1,11 +1,15 @@
 #ifndef INLINEERRORHANDLER_C
 #define INLINEERRORHANDLER_C
     #include <string.h>
+    #include <stdio.h>
+    #include <clist.h>
+    #include <time.h>
 
-    typedef struct  {
-        int error;
-        char* name;
-        char* time;
-    } inlineErrorCode;
-    void logError(inlineErrorCode* code);
+    typedef struct {
+        int code;
+        char* text;
+    } inlineError;
+
+    void logError(int *code, char* name);
+    List getErrList();
 #endif